// frontend/lib/api/generated/model/courseCatalogSearchResult.ts
export interface CourseCatalogSearchResult {
  id: number
  academic_year?: CourseCatalogSearchResultAcademicYear
  section?: CourseCatalogSearchResultSection
  credit_type?: CourseCatalogSearchResultCreditType
  credit_units?: CourseCatalogSearchResultCreditUnits
  domain?: CourseCatalogSearchResultDomain
  course_number?: CourseCatalogSearchResultCourseNumber
  course_code?: CourseCatalogSearchResultCourseCode
  old_course_name?: CourseCatalogSearchResultOldCourseName
  course_name?: CourseCatalogSearchResultCourseName
  course_guide?: CourseCatalogSearchResultCourseGuide
  accreditation_type?: CourseCatalogSearchResultAccreditationType
  degree_program?: CourseCatalogSearchResultDegreeProgram
  instructor?: CourseCatalogSearchResultInstructor
  total_credits?: CourseCatalogSearchResultTotalCredits
  lecture_credits?: CourseCatalogSearchResultLectureCredits
  lab_credits?: CourseCatalogSearchResultLabCredits
  course_category?: CourseCatalogSearchResultCourseCategory
  enrollment_capacity?: CourseCatalogSearchResultEnrollmentCapacity
  class_format?: CourseCatalogSearchResultClassFormat
  class_time?: CourseCatalogSearchResultClassTime
  classroom?: CourseCatalogSearchResultClassroom
  completion_restriction?: CourseCatalogSearchResultCompletionRestriction
  detailed_info?: CourseCatalogSearchResultDetailedInfo
  core_competency?: CourseCatalogSearchResultCoreCompetency
  offering_department?: CourseCatalogSearchResultOfferingDepartment
  course_overview?: CourseCatalogSearchResultCourseOverview
  course_objectives?: CourseCatalogSearchResultCourseObjectives
  week1_plan?: CourseCatalogSearchResultWeek1Plan
  week2_plan?: CourseCatalogSearchResultWeek2Plan
  week3_plan?: CourseCatalogSearchResultWeek3Plan
  week4_plan?: CourseCatalogSearchResultWeek4Plan
  week5_plan?: CourseCatalogSearchResultWeek5Plan
  week6_plan?: CourseCatalogSearchResultWeek6Plan
  week7_plan?: CourseCatalogSearchResultWeek7Plan
  week8_plan?: CourseCatalogSearchResultWeek8Plan
  week9_plan?: CourseCatalogSearchResultWeek9Plan
  week10_plan?: CourseCatalogSearchResultWeek10Plan
  week11_plan?: CourseCatalogSearchResultWeek11Plan
  week12_plan?: CourseCatalogSearchResultWeek12Plan
  week13_plan?: CourseCatalogSearchResultWeek13Plan
  week14_plan?: CourseCatalogSearchResultWeek14Plan
  week15_plan?: CourseCatalogSearchResultWeek15Plan
  week16_plan?: CourseCatalogSearchResultWeek16Plan
}